Mod Compatibility: replace 'hold shift' with setting (default on)
