+++ Source/Client/Multiplayer.cs
@@ -8,8 +8,6 @@ using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-using RestSharp;
 using HarmonyLib;
 using Multiplayer.Common;
 using RimWorld;
@@ -153,23 +151,8 @@ namespace Multiplayer.Client
                 RuntimeHelpers.RunClassConstructor(typeof(Text).TypeHandle);
             }
             else {
-                UpdateModCompatibilityDb();
+                ModManagement.UpdateModCompatibilityDb();
             }
-        }
-
-        private static void UpdateModCompatibilityDb()
-        {
-            Task.Run(() => {
-                var client = new RestClient("https://bot.rimworldmultiplayer.com/mod-compatibility?version=1.1");
-                try {
-                    var rawResponse = client.Get(new RestRequest($"", DataFormat.Json));
-                    modsCompatibility = SimpleJson.DeserializeObject<Dictionary<string, int>>(rawResponse.Content);
-                    Log.Message($"MP: successfully fetched {modsCompatibility.Count} mods compatibility info");
-                }
-                catch (Exception e) {
-                    Log.Warning($"MP: updating mod compatibility list failed {e.Message} {e.StackTrace}");
-                }
-            });
         }
 
         private static void SetUsername()
